refactor: CompStatusButton 상수 분리
